+++ client/src/components/WorkshopHome/WorkshopHome.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Aux from '../hocs/Aux';
-import Workshop from './Workshop/Workshop';
+import WorkshopPreview from './WorkshopPreview/WorkshopPreview';
 import './WorkshopHome.css';
 
 import { gql, useQuery } from '@apollo/client';
@@ -9,16 +9,11 @@ const GET_WORKSHOPS = gql`
 	query GetWorkshops {
 		workshops {
 			id
-			sections {
-				lessons {
-					lessonTitle
-				}
-				sectionTitle
-			}
 			workshopTitle
 			previewPhoto {
 				url
 			}
+			descriptionPreview
 		}
 	}
 `;
@@ -31,13 +26,15 @@ export default function WorkshopHome() {
 	return (
 		<Aux>
 			<h1 id='page-title'>Workshops</h1>
-			{data.workshops.map((workshop, key) => (
-				<Workshop
-					key={workshop.id}
-					workshopTitle={workshop.workshopTitle}
-					workshopPreview={workshop.workshopPreview}
-					workshopImage={workshop.previewPhoto.url}
-				/>
+			{data.workshops.map((workshop) => (
+				<Aux key={workshop.id}>
+					<WorkshopPreview
+						keyVal={workshop.id}
+						workshopTitle={workshop.workshopTitle}
+						workshopPreview={workshop.descriptionPreview}
+						workshopImage={workshop.previewPhoto.url}
+					/>
+				</Aux>
 			))}
 		</Aux>
 	);
