+++ client/src/components/WorkshopHome/WorkshopSections/WorkshopSections.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import Aux from '../../hocs/Aux';
 import { gql, useQuery } from '@apollo/client';
-import WorkshopLesson from './WorkshopLesson/WorkshopLesson';
+import WorkshopLessons from './WorkshopLessons/WorkshopLessons';
 
 const GET_SECTIONS = gql`
 	query GetSections {
 		sections {
 			lessons {
+				id
 				lessonTitle
 			}
 			sectionTitle
@@ -14,23 +15,17 @@ const GET_SECTIONS = gql`
 	}
 `;
 
-export default function WorkshopSection() {
+export default function WorkshopSections() {
 	const { loading, error, data } = useQuery(GET_SECTIONS);
 
 	if (loading) return 'Loading...';
 	if (error) return `Error! ${error.message}`;
 	return (
 		<Aux>
-			<h1>WorkshopSection</h1>
 			{data.sections.map(({ sectionTitle, lessons }, index) => (
 				<div key={index}>
 					<h2>{sectionTitle}</h2>
-					{lessons.map((lesson, indexVal) => (
-						<WorkshopLesson
-							lessonTitle={lesson.lessonTitle}
-							key={indexVal}
-						/>
-					))}
+					<WorkshopLessons lessons={lessons} />
 				</div>
 			))}
 		</Aux>
